single collection layout and batch update
