+++ novawallet/Modules/AssetList/AssetListViewLayout.swift
@@ -3,17 +3,17 @@ import UIKit
 final class AssetListViewLayout: UIView {
     let backgroundView = MultigradientView.background
 
-    let collectionNetworkGroupsLayout = AssetListNetworksFlowLayout()
-    let collectionTokenGroupsLayout = AssetListTokensFlowLayout()
+    let collectionViewLayout: AssetListFlowLayout = {
+        let layout = AssetListFlowLayout()
+        layout.scrollDirection = .vertical
+        layout.minimumLineSpacing = 0
+        layout.minimumInteritemSpacing = 0
+        layout.sectionInset = .zero
+
+        return layout
+    }()
 
     var assetGroupsLayoutStyle: AssetListGroupsStyle?
-
-    var collectionViewLayout: AssetListFlowLayout {
-        switch assetGroupsLayoutStyle ?? .tokens {
-        case .networks: collectionNetworkGroupsLayout
-        case .tokens: collectionTokenGroupsLayout
-        }
-    }
 
     lazy var collectionView: UICollectionView = {
         let view = UICollectionView(
@@ -40,8 +40,6 @@ final class AssetListViewLayout: UIView {
     }
 
     func setup() {
-        setupLayouts()
-
         addSubview(backgroundView)
         backgroundView.snp.makeConstraints { make in
             make.edges.equalToSuperview()
@@ -50,18 +48,6 @@ final class AssetListViewLayout: UIView {
         addSubview(collectionView)
         collectionView.snp.makeConstraints { make in
             make.edges.equalToSuperview()
-        }
-    }
-
-    func setupLayouts() {
-        [
-            collectionNetworkGroupsLayout,
-            collectionTokenGroupsLayout
-        ].forEach {
-            $0.scrollDirection = .vertical
-            $0.minimumLineSpacing = 0
-            $0.minimumInteritemSpacing = 0
-            $0.sectionInset = .zero
         }
     }
 }
